refactor graph info menu
