+++ client/src/products/graph-sandbox/ui/GraphInfoMenu/AdjacencyListDisplay.vue
