+++ client/src/products/graph-sandbox/ui/GraphInfoMenu/TransitionMatrixLabel.vue
